backend/requirements.txt
# requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
requests>=2.25.0
firebase-admin>=5.0.0
pyjwt>=2.0.0
huggingface-hub>=0.10.0
numpy
sentencepiece
scikit-learn
